+++ src/app/Components/ComponentsSamples/Lists/ScrollViewSamples.xaml.cs
@@ -1,9 +1,0 @@
-namespace Components.ComponentsSamples.Lists;
-
-public partial class ScrollViewSamples
-{
-    public ScrollViewSamples()
-    {
-        InitializeComponent();
-    }
-}